@@ -1,16 +1,16 @@
 ﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 
-namespace GraphQL.AspNetCore.Graphiql
+namespace GraphQL.AspNetCore.GraphiQL
 {
-    public class GraphiqlMiddleware
+    public class GraphiQLMiddleware
     {
         private readonly RequestDelegate _next;
-        private readonly GraphiqlMiddlewareOptions _options;
+        private readonly GraphiQLMiddlewareOptions _options;
 
-        public GraphiqlMiddleware(
+        public GraphiQLMiddleware(
             RequestDelegate next,
-            GraphiqlMiddlewareOptions options)
+            GraphiQLMiddlewareOptions options)
         {
             _next = next;
             _options = options;
@@ -39,15 +39,14 @@ namespace GraphQL.AspNetCore.Graphiql
         }
         #graphiql {
             height: 98vh;
-
+        }
         .jwt-token {
-                background: linear-gradient(#f7f7f7, #e2e2e2);
-                border-bottom: 1px solid #d0d0d0;
-                font-family: system, -apple-system, 'San Francisco', '.SFNSDisplay-Regular', 'Segoe UI', Segoe, 'Segoe WP', 'Helvetica Neue', helvetica, 'Lucida Grande', arial, sans-serif;
-                padding: 7px 14px 6px;
-                font-size: 14px;
-              }
-      }
+            background: linear-gradient(#f7f7f7, #e2e2e2);
+            border-bottom: 1px solid #d0d0d0;
+            font-family: system, -apple-system, 'San Francisco', '.SFNSDisplay-Regular', 'Segoe UI', Segoe, 'Segoe WP', 'Helvetica Neue', helvetica, 'Lucida Grande', arial, sans-serif;
+            padding: 7px 14px 6px;
+            font-size: 14px;
+        }
     </style>
     <link href=""//unpkg.com/graphiql@0.11.2/graphiql.css"" rel=""stylesheet"">
     <script src=""//unpkg.com/react@15.6.1/dist/react.min.js""></script>
